+++ examples/Direct/sendText.php
@@ -62,6 +62,7 @@ try {
         ],
     ];
 
+    $ig->direct->fetchAndSubscribePresence($userId);
     $ig->direct->getThreadByParticipants([$userId]);
     $ig->direct->sendText($recipients, $text, ['client_context' => $clientContext, 'send_attribution' => 'inbox_new_message']);
     $ig->event->sendDirectMessageIntentOrAttempt('send_intent', $clientContext, 'text', [$userId]);